requirements.txt
streamlit
matplotlib
google-genai>=1.0.0
langchain-google-genai>=2.1.0
langgraph>=0.0.30
langchain>=0.1.0
# RAG Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf2>=3.0.0
python-docx>=0.8.11
langchain-community>=0.0.10
langchain-chroma>=0.1.0
tiktoken>=0.5.0